Made base webpack config take a packagePath parameter and moved it to the monorepo root
