+++ packages/api-explorer/webpack.dev.config.js
@@ -25,7 +25,7 @@
  */
 
 const path = require('path')
-const base = require('./webpack.base.config')
+const base = require('../../webpack.base.config')(__dirname)
 
 module.exports = {
   ...base,
