+++ packages/api-explorer/webpack.prod.config.js
@@ -23,7 +23,7 @@
  SOFTWARE.
 
  */
-const base = require('./webpack.base.config')
+const base = require('../../webpack.base.config')(__dirname)
 
 module.exports = {
   ...base,
